Refactor: DashboardSidebar into smaller components
Refactor DashboardSidebar.tsx into smaller, reusable components for better organization and maintainability.
